RN-699: Move AccessPolicy into its own file
